create new file for SpanExpressionHelper class
